合并

+++ src/lvio_fusion/include/lvio_fusion/utility.h
@@ -33,10 +33,10 @@ inline void triangulate(const SE3d &pose0, const SE3d &pose1, const Vector3d &p0
     p_3d = (p_norm / p_norm(3)).head<3>();
 }
 
-inline Vector2d cv2eigen(const cv::Point2d &p) { return Vector2d(p.x, p.y); }
-inline Vector3d cv2eigen(const cv::Point3d &p) { return Vector3d(p.x, p.y, p.z); }
-inline cv::Point2d eigen2cv(const Vector2d &p) { return cv::Point2d(p.x(), p.y()); }
-inline cv::Point3d eigen2cv(const Vector3d &p) { return cv::Point3d(p.x(), p.y(), p.z()); }
+inline Vector2d cv2eigen(const cv::Point2f &p) { return Vector2d(p.x, p.y); }
+inline Vector3d cv2eigen(const cv::Point3f &p) { return Vector3d(p.x, p.y, p.z); }
+inline cv::Point2f eigen2cv(const Vector2d &p) { return cv::Point2f(p.x(), p.y()); }
+inline cv::Point3f eigen2cv(const Vector3d &p) { return cv::Point3f(p.x(), p.y(), p.z()); }
 
 /**
  * line fitting
@@ -256,7 +256,6 @@ inline T normalize_angle(const T &angle_degrees)
                two_pi * std::floor((-angle_degrees + T(180)) / two_pi);
 };
 
-// NEWADD
 inline Eigen::Vector3d LogSO3(const Eigen::Matrix3d &R)
 {
     const double tr = R(0,0)+R(1,1)+R(2,2);
@@ -394,7 +393,8 @@ inline Eigen::Matrix3d RightJacobianSO3(const Eigen::Vector3d &v)
     return RightJacobianSO3(v[0],v[1],v[2]);
 }
 
-//NEWADDEND
+
+
 } // namespace lvio_fusion
 
 #endif // lvio_fusion_UTILITY_H
