+++ src/lvio_fusion/include/lvio_fusion/visual/feature.h
@@ -20,7 +20,7 @@ public:
 
     Feature() {}
 
-    static Feature::Ptr Create(std::shared_ptr<Frame> frame, const cv::Point2d &kp, std::shared_ptr<Landmark> landmark)
+    static Feature::Ptr Create(std::shared_ptr<Frame> frame, const cv::Point2f &kp, std::shared_ptr<Landmark> landmark)
     {
         Feature::Ptr new_feature(new Feature);
         new_feature->frame = frame;
@@ -30,7 +30,7 @@ public:
     }
 
     std::weak_ptr<Frame> frame;
-    cv::Point2d keypoint;
+    cv::Point2f keypoint;
     std::weak_ptr<Landmark> landmark;
     bool is_on_left_image = true;
 };
